feat(admin): add modular product management page with category filter and delete functionality
