+++ handcrafted-heaven-frontend/app/admin/dashboard/users/page.tsx
@@ -1,27 +1,15 @@
 "use client";
 
-import { useEffect, useState } from "react";
-import { useRouter } from "next/navigation";
+import { useState, useEffect } from "react";
 import api from "@/lib/api";
-import { UserRoundCog } from "lucide-react";
-
-type User = {
-  id: string;
-  name: string;
-  email: string;
-  role: string;
-  createdAt: string;
-  orders: { id: string }[];
-  seller: { id: string } | null;
-  Image: { url: string } | null;
-};
+import UserTable from "@/components/admin/UserTable";
+import Pagination from "@/components/admin/Pagination";
 
 export default function AdminUsersPage() {
-  const [users, setUsers] = useState<User[]>([]);
+  const [users, setUsers] = useState([]);
   const [loading, setLoading] = useState(true);
   const [page, setPage] = useState(1);
   const [totalPages, setTotalPages] = useState(1);
-  const router = useRouter();
 
   useEffect(() => {
     setLoading(true);
@@ -38,84 +26,15 @@ export default function AdminUsersPage() {
   return (
     <div className="p-6">
       <h1 className="text-2xl font-semibold mb-6">Users</h1>
-
       {loading ? (
         <p className="text-center text-gray-500">Loading users...</p>
       ) : (
         <>
-          <div className="overflow-x-auto">
-            <table className="min-w-full bg-white shadow rounded-md">
-              <thead className="bg-gray-100 text-left">
-                <tr>
-                  <th className="px-4 py-2">Name</th>
-                  <th className="px-4 py-2">Email</th>
-                  <th className="px-4 py-2">Role</th>
-                  <th className="px-4 py-2">Orders</th>
-                  <th className="px-4 py-2">Type</th>
-                  <th className="px-4 py-2">Joined</th>
-                  <th className="px-4 py-2">Actions</th>
-                </tr>
-              </thead>
-              <tbody>
-                {users.map((user) => (
-                  <tr key={user.id} className="border-t">
-                    <td className="px-4 py-2 flex items-center gap-2">
-                      {user.Image?.url && (
-                        <img
-                          src={user.Image.url}
-                          alt={user.name}
-                          className="w-8 h-8 rounded-full"
-                        />
-                      )}
-                      {user.name}
-                    </td>
-                    <td className="px-4 py-2">{user.email}</td>
-                    <td className="px-4 py-2">{user.role}</td>
-                    <td className="px-4 py-2">{user.orders.length}</td>
-                    <td className="px-4 py-2">
-                      {user.seller ? "Seller" : "Customer"}
-                    </td>
-                    <td className="px-4 py-2">
-                      {new Date(user.createdAt).toLocaleDateString()}
-                    </td>
-                    <td className="px-4 py-2">
-                      <button
-                        onClick={() =>
-                          router.push(`/admin/dashboard/user/${user.id}`)
-                        }
-                        className="flex items-center gap-1 text-blue-600 hover:underline"
-                      >
-                        <UserRoundCog size={18} />
-                        Manage
-                      </button>
-                    </td>
-                  </tr>
-                ))}
-              </tbody>
-            </table>
-          </div>
-
-          <div className="flex justify-center mt-4 gap-2">
-            <button
-              disabled={page === 1}
-              onClick={() => setPage((p) => p - 1)}
-              className="px-3 py-1 bg-gray-200 rounded disabled:opacity-50"
-            >
-              Prev
-            </button>
-            <span className="px-3 py-1 font-medium">
-              Page {page} of {totalPages}
-            </span>
-            <button
-              disabled={page === totalPages}
-              onClick={() => setPage((p) => p + 1)}
-              className="px-3 py-1 bg-gray-200 rounded disabled:opacity-50"
-            >
-              Next
-            </button>
-          </div>
+          <UserTable users={users} />
+          <Pagination page={page} totalPages={totalPages} onPageChange={setPage} />
         </>
       )}
     </div>
   );
 }
+
